+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 # @oslojs/crypto
+
+## 0.2.1
+
+- Add `@oslojs/crypto/ecdsa`
+- Export `sha224()` and `SHA224`